Refactor command_handler
