+++ mqtt/mqtt-edgehub/tests/command.rs
@@ -1,8 +1,8 @@
-use std::time::Duration;
+use std::{collections::HashMap};
 
 use futures_util::StreamExt;
 use mqtt3::{proto::ClientId, ShutdownError};
-use tokio::{task::JoinHandle, time};
+use tokio::{task::JoinHandle};
 
 use mqtt_broker::{auth::AllowAll, BrokerBuilder, BrokerHandle};
 use mqtt_broker_tests_util::{
@@ -10,8 +10,14 @@ use mqtt_broker_tests_util::{
     packet_stream::PacketStream,
     server::{start_server, DummyAuthenticator},
 };
-use mqtt_edgehub::command::{CommandHandler, ShutdownHandle};
+use mqtt_edgehub::command::{
+    handle_authorized_identities, handle_disconnect, Command,
+};
 
+use mqtt_edgehub::command_handler::{ CommandHandler, ShutdownHandle };
+
+const DISCONNECT_TOPIC: &str = "$edgehub/disconnect";
+const AUTHORIZED_IDENTITIES_TOPIC: &str = "$internal/identities";
 const TEST_SERVER_ADDRESS: &str = "localhost:5555";
 
 /// Scenario:
@@ -27,8 +33,9 @@ async fn disconnect_client() {
 
     let server_handle = start_server(broker, DummyAuthenticator::anonymous());
 
+    let commands = init_commands();
     let (command_handler_shutdown_handle, join_handle) =
-        start_command_handler(broker_handle, TEST_SERVER_ADDRESS.to_string())
+        start_command_handler(broker_handle, TEST_SERVER_ADDRESS.to_string(), commands)
             .await
             .expect("could not start command handler");
 
@@ -45,7 +52,7 @@ async fn disconnect_client() {
         .with_client_id(ClientId::IdWithCleanSession("$edgehub".into()))
         .build();
 
-    let topic = "$edgehub/disconnect";
+    let topic = DISCONNECT_TOPIC;
     edgehub_client
         .publish_qos1(topic, r#""test-client""#, false)
         .await;
@@ -62,12 +69,32 @@ async fn disconnect_client() {
     edgehub_client.shutdown().await;
 }
 
+fn init_commands() -> HashMap<String, Command> {
+    let mut commands = HashMap::new();
+    commands.insert(
+        DISCONNECT_TOPIC.to_string(),
+        Command {
+            topic: DISCONNECT_TOPIC.to_string(),
+            handle: handle_disconnect,
+        },
+    );
+    commands.insert(
+        AUTHORIZED_IDENTITIES_TOPIC.to_string(),
+        Command {
+            topic: AUTHORIZED_IDENTITIES_TOPIC.to_string(),
+            handle: handle_authorized_identities,
+        },
+    );
+    commands
+}
+
 async fn start_command_handler(
     broker_handle: BrokerHandle,
     system_address: String,
+    commands: HashMap<String, Command>,
 ) -> Result<(ShutdownHandle, JoinHandle<()>), ShutdownError> {
     let device_id = "test-device";
-    let command_handler = CommandHandler::new(broker_handle, system_address, device_id)
+    let command_handler = CommandHandler::new(broker_handle, system_address, device_id, commands)
         .await
         .unwrap();
     let shutdown_handle: ShutdownHandle = command_handler.shutdown_handle().unwrap();
